completed the functions refactor
